using same prompt and full precision (off by 1%)

+++ examples/llm/tech_qa.py
@@ -237,7 +237,7 @@ def train(args, data_lists):
     gnn = GAT(in_channels=768, hidden_channels=hidden_channels,
               out_channels=1024, num_layers=num_gnn_layers, heads=4)
     # freeze LM
-    llm = LLM(model_name=args.llm_generator_name).eval()
+    llm = LLM(model_name=args.llm_generator_name, dtype=torch.float).eval()
     #
     return llm, DataLoader(
         data_lists["train"] + data_lists["validation"] + data_lists["test"],
